Parallelizing the global registration, issue #217. The global registration had to be ported to the generic processing function to do so, which wasn't straightforward: - image hook has the same arguments as save hook now, with the two indices passed - improved the output files of the generic sequence processing with the actual file number instead of the index of the input image - added a thread-safe redraw in case we want to draw stars from the registration - adding a crazy SER compacting function that allows SER to be created with holes during a parallel processing that can partially fail - the loading of the new sequence after registration is now managed by the generic registration idle and not by the global registration - moved all that to a new registration/global.c file
Registration now sets the best image as reference image of the sequence if none was defined.

Automatic exclusion of partial failures in generic sequence processing has been
disabled as it caused too many issues and was not functionally great imho.

Identifying the error condition in stats too, but not fixing it. Changing the way star_finder_params is used.
Saving stats at the end of a generic sequence processing because some may have been generated and they were not cached.