Use system shared librtprocess if available.
The configure.ac will search for presence of librtprocess through
pkg-config. If absent **and** if not in a cross-compilation environment,
then sets up building it with the internal submodule source and CMake.

Note that when cross-building, it is not easily feasible to build it
(unless you add some dependency doing proper set-up of cross-compilation
options for you such as crossroad) so I just forbid it in such case and
only allow internal static compilation for simpler native builds.

Also note that configuration of a submodule must not happen in autogen
(which makes it in-build and once and for all for all builds, without
any basic check of any sort). If you really want this to be done this
way, it must be in the automake files.