Add the CMake build system generator option back for Windows.
This option was in the autogen.sh (based on uname instead of proper
detection by configure script, which is now also changed). Add it back
properly in the Makefile rule. Of course it will be used only for
**native** builds on Windows for Windows. Cross-builds for the
librtprocess dependency are still to be done separately.

Also move the rules inside deps/Makefile as it should be, now that we
have a Makefile.am there.