Improve gitlab-ci.
Let's make an out-of-tree build into the _build/ directory. Install
prefix is now done in ${INSTALL_PREFIX} (note that most of the
environment variables defined previously were useless as they were based
on INSTALL_PREFIX but nothing was installed there!).

Finally save both the prefix **and** the build results. Otherwise there
is no binary left for debugging if we don't reach install step (which is
most broken cases when a bug appear in the CI).