deps: use proper pkg-config result in kplot for cross-compilation.
Using directly `pkg-config` is definitely not right as this is the tool
to find native libraries. We could use $(PKG_CONFIG) (which is the
properly detected pkg-config targetted to the host) but the better way
is to have real dependency check in the configure step, and use the
result _LIBS and _CFLAGS variable in the Makefile.

The fact it even worked before in some cases when cross-compiling was
just pure luck (basically it was getting the native Cairo header file
which happened to be OK, but that is definitely not proper)!