When drawing a selection, display a 3 (or 5) thirds grid inside the rectangle.
This grid disappears when the selection is done, and reappears when the selection is modified by the user.

Useful when making a big selection for cropping the image, either for centering purpose, or for composition:
The rule of thirds is one of the most useful composition techniques in photography.

Simple "Quality Of Life" improvement in a series of improvements I would like to do to the selection tool.

- Also increase the width of the selection from 0.8 to 1.5 so that it's more visible
- This grid is fainter, using 0.5 width for now